adjust Before-specific v1/games logic
Before now asks Chronicler for the list of games for a season/day pair
to get start times to jump to dynamically, instead of querying v1/games
for all games and creating a cache.

This fixes the Before-specific logic to filter the list of games
returned based on the season and day query parameters.